Moved detr loss to loss folder, added detr_sfd class, minor pep-8 compliance changes.
