+++ nnet_training/loss_functions/__init__.py
@@ -10,6 +10,7 @@ from .depth_losses import InvHuberLoss, InvHuberLossPyr, DepthAwareLoss, ScaleIn
 from .UnFlowLoss import unFlowLoss
 from .rmi import RMILoss, RMILossAux, MultiScaleRMILoss
 from .seg_losses import FocalLoss2D, SegCrossEntropy
+from .detr import DetrLoss
 
 def get_loss_function(loss_config) -> Dict[str, torch.nn.Module]:
     """
@@ -38,6 +39,8 @@ def get_loss_function(loss_config) -> Dict[str, torch.nn.Module]:
             loss_fn_dict[loss_fn['type']] = MultiScaleRMILoss(**loss_fn.args)
         elif loss_fn['function'] == "SegCrossEntropy":
             loss_fn_dict[loss_fn['type']] = SegCrossEntropy(**loss_fn.args)
+        elif loss_fn['function'] == "DetrLoss":
+            loss_fn_dict[loss_fn['type']] = DetrLoss(**loss_fn.args)
         else:
             raise NotImplementedError(loss_fn['function'])
